% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_deps.R
\name{audit_deps}
\alias{audit_deps}
\title{Check Package Dependencies

Collects R dependencies and checks them against OSS Index.
Returns a tibble of results.}
\usage{
audit_deps()
}
\description{
Check Package Dependencies

Collects R dependencies and checks them against OSS Index.
Returns a tibble of results.
}
