% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{audit_deps_with_oss_index}
\alias{audit_deps_with_oss_index}
\title{Convenient function that collects dependencies, checks them against OSS Index, and audits them}
\usage{
audit_deps_with_oss_index()
}
\description{
Convenient function that collects dependencies, checks them against OSS Index, and audits them
}
