% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_audits.R
\name{audit_description}
\alias{audit_description}
\title{Audits Packages Listed in a DESCRIPTION file}
\usage{
audit_description(
  dir = ".",
  fields = c("Depends", "Imports", "Suggests"),
  verbose = TRUE
)
}
\arguments{
\item{dir}{The file path of an renv.lock file.}

\item{fields}{The DESCRIPTION field to parse. Default is Depends, Import, & Suggests.}

\item{verbose}{Default \code{TRUE}.}
}
\description{
Looks for a DESCRIPTION file in \code{dir}, then extract
the packages in the fields & calculates the dependency tree.
}
\examples{
\dontrun{
# Looks for a DESCRIPTION file in dir
audit_description(dir = ".")
}
}
