% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_audits.R
\name{audit_env_yml}
\alias{audit_env_yml}
\title{Audit a conda environment file}
\usage{
audit_env_yml(dir = ".", fname = "environment.yml", verbose = TRUE)
}
\arguments{
\item{dir}{The directory containing a conda environment yaml file.}

\item{fname}{The file name of conda environment yaml file.}

\item{verbose}{Default \code{TRUE}.}
}
\description{
This function searches the OSS index for vulnerabilities recorded for packages listed
in a Conda environment file typically called environment.yml but are subject to varied names.
Conda environment can contain packages from both Conda and PyPI. All packages will be audited.
}
\examples{
\dontrun{
# Looks for a requirements.txt file in dir
audit_env_yml(dir = ".")
}
}
