% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{audit_installed_r_pkgs}
\alias{audit_installed_r_pkgs}
\title{Audit Installed Packages}
\usage{
audit_installed_r_pkgs(verbose = TRUE, token = NULL)
}
\arguments{
\item{verbose}{Default \code{TRUE}.}

\item{token}{If NULL, looks at OSSINDEX_USER & OSSINDEX_TOKEN, env variables. If those
aren't available, try \code{"~/.ossindex/.oss-index-config"}}
}
\value{
A tibble/data.frame.
}
\description{
Audits all installed packages by calling \code{installed.packages()}
and checking them against the OSS Index.
}
\examples{
\dontrun{
# Audit installed packages
# This calls installed.packages()
pkgs = audit_installed_r_pkgs()
}
}
