% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_deps.R
\name{audit_pkgs}
\alias{audit_pkgs}
\title{Search for package vulnerabilities}
\usage{
audit_pkgs(pkg, version = "*", type = "cran", verbose = TRUE)
}
\arguments{
\item{pkg}{A vector of package names to search in the OSS Index.}

\item{version}{The specific package version to search for.
By default it will search all known versions. If not \code{*}, must be the same length as pkg.}

\item{type}{The package management environment.
This defaults to \code{"cran"}. See https://ossindex.sonatype.org/ecosystems.}

\item{verbose}{Default \code{TRUE}.}
}
\description{
Search the OSS Index for known package vulnerabilities in any of the supported ecosystems—
e.g. CRAN, PyPI, Conda, NPM, Maven, etc.
see https://ossindex.sonatype.org/ecosystems for full list.
}
