% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_deps.R
\name{audit_pypi_deps}
\alias{audit_pypi_deps}
\title{Check PyPi package dependencies}
\usage{
audit_pypi_deps(pkg, version = "*", verbose = TRUE)
}
\arguments{
\item{pkg}{Name of package as listed on PyPi.}

\item{version}{The version of the package. Optional.}

\item{verbose}{Default \code{TRUE}}
}
\description{
Search the OSS Index for known PyPi software dependency vulnerabilities.
}
\details{
By default \code{audit_pypi_deps()} will return audits for all known versions of a package on PyPi. To search for a single package version, supply the version as a character to \code{version} argument.
}
