% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_audits.R
\name{audit_req_txt}
\alias{audit_req_txt}
\title{Audit a requirements.txt File}
\usage{
audit_req_txt(dir = ".", verbose = TRUE)
}
\arguments{
\item{dir}{The file path of a requirements.txt file.}

\item{verbose}{Default \code{TRUE}.}
}
\description{
This function searches the OSS index for vulnerabilities recorded for packages listed
in a requirements.txt file based on PyPi.
}
\details{
pip is a standard of python package management based on the Python Package Index (PyPI).
pip uses a requirements.txt file to manage of Python libraries.
The requirements.txt file contains package names and versions
(often used to manage a virtual environment).
}
\examples{
\dontrun{
# Looks for a requirements.txt file in dir
audit_description(dir = ".")
}
}
