% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{audit_response_from_oss_index}
\alias{audit_response_from_oss_index}
\title{Audits a response from OSS Index and prints out a software bill of materials}
\usage{
audit_response_from_oss_index(response)
}
\arguments{
\item{response}{A list of coordinates and vulnerabilities from OSS Index}
}
\description{
Audits a response from OSS Index and prints out a software bill of materials
}
