% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{call_oss_index}
\alias{call_oss_index}
\title{Calls OSS Index given a list of purls}
\usage{
call_oss_index(allPurls)
}
\arguments{
\item{allPurls}{A list of purls}
}
\value{
A list of results from OSS Index
}
\description{
Calls OSS Index given a list of purls
}
