% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{collect_dependencies_and_turn_into_purls}
\alias{collect_dependencies_and_turn_into_purls}
\title{Collects dependencies installed locally, and turns them into purls}
\usage{
collect_dependencies_and_turn_into_purls()
}
\value{
A list of purls
}
\description{
Collects dependencies installed locally, and turns them into purls
}
