% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_purls}
\alias{gen_purls}
\title{Function to generate purls}
\usage{
gen_purls(pkg, version = "*", schema = "cran")
}
\description{
Generates purls from a vector of package names, version, and schema. \code{version} and \code{schema} must be the same length as \code{pkg} or else be of length one.
}
\keyword{internal}
