% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_deps.R
\name{get_vulnerabilies}
\alias{get_vulnerabilies}
\title{Extract Vulnerabilities

Parse the audit data frame, and extract
the vulnerabilities.}
\usage{
get_vulnerabilies(audit)
}
\arguments{
\item{audit}{Output from \code{audit_deps}.}
}
\description{
Extract Vulnerabilities

Parse the audit data frame, and extract
the vulnerabilities.
}
