% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{MAX_NUMBER_OF_PURLS_PER_API_CALL}
\alias{MAX_NUMBER_OF_PURLS_PER_API_CALL}
\title{OSS Index max number of purls to use}
\format{An object of class \code{numeric} of length 1.}
\usage{
MAX_NUMBER_OF_PURLS_PER_API_CALL
}
\description{
OSS Index max number of purls to use
}
\keyword{datasets}
