% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{OSS_INDEX_URL}
\alias{OSS_INDEX_URL}
\title{OSS Index public URL}
\format{An object of class \code{character} of length 1.}
\usage{
OSS_INDEX_URL
}
\description{
OSS Index public URL
}
\keyword{datasets}
