% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{audit_deps}
\alias{audit_deps}
\title{Check Package Dependencies}
\usage{
audit_deps(pkgs = NULL, verbose = TRUE)
}
\arguments{
\item{pkgs}{Default \code{NULL}. See details for further information.}

\item{verbose}{Default \code{TRUE}.}
}
\value{
A tibble/data.frame.
}
\description{
Collects R dependencies and checks them against OSS Index.
Returns a tibble of results.
}
\details{
This function is deprecated. See

By default, packages listed in \code{installed.packages()} are scanned by sonatype.
However, you can pass your own data frame of packages. This data frame should have two columns,
\code{version} and \code{package}.
}
