% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{audit_installed_r_pkgs}
\alias{audit_installed_r_pkgs}
\title{Audit Installed Packages}
\usage{
audit_installed_r_pkgs(verbose = TRUE)
}
\arguments{
\item{verbose}{Default \code{TRUE}.}
}
\value{
A tibble/data.frame.
}
\description{
Audits all installed packages by calling \code{installed.packages()}
and checking them against the OSS Index.
}
\examples{
\donttest{
# Audit installed packages
# This calls installed.packages()
pkgs = audit_installed_r_pkgs()
}
}
