% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_audits.R
\name{audit_renv_lock}
\alias{audit_renv_lock}
\title{Audit an renv.lock File}
\usage{
audit_renv_lock(dir = ".", verbose = TRUE)
}
\arguments{
\item{dir}{The file path of an renv.lock file.}

\item{verbose}{Default \code{TRUE}.}
}
\description{
This function searches the OSS index for vulnerabilities recorded for packages listed in
an \code{renv.lock} file.
An \code{renv.lock} file is created by the \code{{renv}} package
which is used for project level package management in R.
}
\examples{
\dontrun{
# Looks for renv.lock file in dir
audit_renv_lock(dir = ".")
}
}
