% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_secure.R
\name{expect_secure}
\alias{expect_secure}
\title{Vulnerability Detection via Testthat}
\usage{
expect_secure(pkg, repo = "https://cran.rstudio.com", verbose = FALSE)
}
\arguments{
\item{pkg}{The pkg to check}

\item{repo}{The CRAN repository, used to get version numbers}

\item{verbose}{Default \code{TRUE}.}
}
\description{
A \code{testthat} version for detecting vulnerabilities.
This function is used within the \code{testthat} framework.
As testthat strips out the repositories from options,
we have to set the value locally in the function, i.e. the
value you have in \code{getOption("repos")} is not used.
}
\details{
An important proviso is that we are only testing packages for specific versions.
By default, this will be the latest version on CRAN.
This may differ for users or if you are using a CRAN snapshot.
For the latter, simply change the \code{repo} parameter.
}
\examples{
\dontrun{
 # Typically used inside testthat
 oysteR::expect_secure("oysteR")
}
}
