% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_purls}
\alias{generate_purls}
\title{Function to generate purls}
\usage{
generate_purls(pkg, version, type)
}
\description{
Generates purls from a vector of package names, version, and type. \code{version}
must be the same length as \code{pkg}.
\code{type} must of the same length or else be of length one.
}
\keyword{internal}
