% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_deps.R
\name{get_vulnerabilities}
\alias{get_vulnerabilities}
\title{Extract vulnerabilities

Parse the audit data frame, and extract
the vulnerabilities.}
\usage{
get_vulnerabilities(audit)
}
\arguments{
\item{audit}{Output from \code{audit_deps}.}
}
\description{
Extract vulnerabilities

Parse the audit data frame, and extract
the vulnerabilities.
}
